chore(video): correctness and hygiene cleanups in playback layer
Round-up of the small leftovers from the audit. None move the needle on
their own; together they remove a real cancellation bug and tighten the
playback types.

- PlaybackServiceClient.executorService: Executors.newCachedThreadPool()
  → Executors.newSingleThreadExecutor(). The work per callback is
  Future.get() on an already-completed future plus a non-blocking
  trySend; a single thread is plenty. The previous unbounded pool could
  spin up a thread per concurrent video, each lingering for the 60 s
  keep-alive afterwards.

- MediaControllerState.controller: var → val. The field was never
  reassigned anywhere (grep confirms), and a non-observable var on a
  @Stable class is a footgun — Compose can't see writes to a plain var,
  so any future write would silently miss recomposition.

- MediaControllerState.currrentMedia() → currentMedia(). Typo. Updated
  the single caller in PipVideoView.

- LoadThumbAndThenVideoView: real cancellation bug fix. The Coil fetch
  was launched into AccountViewModel.viewModelScope via a side helper
  (loadThumb), so a scroll-away didn't cancel the in-flight image
  request — wasted bandwidth and a late callback writing into stale
  state. Inline the Coil call into the LaunchedEffect's own scope so
  cancellation propagates, and key the effect on thumbUri so a recycled
  audio-track slot with a new cover doesn't stall on the prior
  Pair(true, ...) gate. Drop the now-unused AccountViewModel.loadThumb
  and its only-here imports.
